+++ DAL/DBMODEL/Emplyee.cs
@@ -32,7 +32,7 @@ namespace DAL.DBMODEL
         public DateTime? Createdon { get; set; }
         public string Password { get; set; }
         public bool? IsLoginAllowed { get; set; }
-        public long? Salary { get; set; }
+        public decimal? Salary { get; set; }
         public string SalaryType { get; set; }
         public decimal? WorkingHours { get; set; }
         public string Image { get; set; }
